move apply_changes to shard tries
