+++ runtime/runtime/src/genesis.rs
@@ -99,12 +99,12 @@ impl GenesisStateApplier {
         let (mut store_update, new_state_root) = tries.apply_all(&trie_changes, shard_uid);
         #[cfg(not(feature = "protocol_feature_flat_state"))]
         let (store_update, new_state_root) = tries.apply_all(&trie_changes, shard_uid);
-        for change in state_changes.drain(..) {
+
+        {
             #[cfg(feature = "protocol_feature_flat_state")]
-            store_update.apply_change_to_flat_state(&change);
-            #[cfg(not(feature = "protocol_feature_flat_state"))]
-            let _ = change;
+            tries.apply_changes_to_flat_state(&state_changes, &mut store_update);
         }
+
         store_update.commit().expect("Store update failed on genesis initialization");
         *current_state_root = new_state_root;
     }
